Added initial support for ModSecurity recording
I had to change the hook from post read to fixup to catch
the headers set by mod_security. This demonstrates that I can catch
them in route and respond to them. This will help track and create
a dashboard of WAF related activity, as well as potentially decide
to trigger action against the repsheet.

This is still very much a work in progress. I probably will end up
storing more information on the waf activity, perhaps potentially
creating a list record for each rule triggered with the corresponding
IP as an item of the list.

Signed-off-by: Aaron Bedra <abedra@modsecurity.(none)>